requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.10.5
PyYAML==6.0.3
numpy==2.4.2
pandas==3.0.0
opencv-python==4.13.0.92
matplotlib==3.10.8
reportlab==4.2.5
ultralytics==8.3.40
torch==2.10.0
torchvision==0.25.0
